+++ lib/async.dart
@@ -8,7 +8,10 @@ library rikulo_async;
 import "dart:async";
 //import "dart:collection";
 
+import "util.dart";
+
 export "src/async/defer.dart";
 
 part "src/async/stream_provider.dart";
 part "src/async/stream_wrapper.dart";
+part "src/async/streams.dart";