+++ lib/util.dart
@@ -4,6 +4,7 @@ library rikulo_util;
 
 //don't include dart:html since this library is designed to work
 //at both client and server
+import "dart:async";
 import "dart:math" show max, min;
 import "dart:collection";
 import "package:charcode/ascii.dart";
@@ -13,6 +14,7 @@ part "src/util/xmls.dart";
 part "src/util/lists.dart";
 part "src/util/maps.dart";
 part "src/util/color.dart";
+part "src/util/invokes.dart";
 
 /** A function that returns an integer.
  */